Encoder connection message (8.0-dev/LTS): Speak and/or braille encoder connection message a few seconds after connecting. re #5.
Requested by a broadcaster: there might be times when braille users would like to get a confirmation that connection has been established. For now, speech and/or braille will be sent (via ui.message/Encoder.encoderStatusMessage function), with the default delay being 3 seconds.
If it's accepted, it'll become part of add-on 8.0 and LTS.
P.S. Starting with this branch, issues will be refer3enced (GitHub) and a new branch name format (version/iNumberTopic) will be used.